4 таблицы

+++ db/migrate/20221207200512_create_abonents.rb
@@ -2,7 +2,7 @@ class CreateAbonents < ActiveRecord::Migration[7.0]
   def change
     create_table :abonents do |t|
       t.string :name
-      t.integer :passport
+      t.bigint :passport
       t.string :address, array: true
 
       t.timestamps
